+++ src/components/Alert/AlertError.tsx
@@ -1,9 +1,9 @@
 function AlertError() {
   return (
-    <div className="bg-yellow-50 h-12 w-full border-orange-200 border rounded-sm pl-4 content-center flex items-center text-orange-800">
+    <div className="bg-yellow-50 h-12 w-full border-orange-200 border rounded-sm pl-4 content-center tw-flex items-center text-orange-800">
       Đã có lỗi xảy ra, vui lòng quay lại sau!
     </div>
-  )
+  );
 }
 
 export default AlertError;