+++ src/components/Layout/MainLayout.tsx
@@ -5,9 +5,9 @@ import Footer from '../Footer';
 
 export default function MainLayout({ children }: LayoutPropsInterface) {
   return (
-    <div className="bg-gray-100 min-h-screen font-serif">
+    <div className="tw-bg-gray-100 tw-min-h-screen tw-font-serif">
       <Header />
-      <div className="max-w-4xl my-0 mx-auto px-4">{children}</div>
+      <div className="tw-max-w-4xl tw-my-0 tw-mx-auto tw-px-4">{children}</div>
       <Footer />
     </div>
   );