+++ src/pages/book/index.tsx
@@ -28,7 +28,7 @@ function BookPage(props: any) {
   const data = props as ArticleListInterface;
 
   return (
-    <div className="grid gap-4 justify-items-center md:grid-cols-2 grid-cols-1 mt-8 mb-4">
+    <div className="tw-grid tw-gap-4 tw-justify-items-center md:tw-grid-cols-2 tw-grid-cols-1 tw-mt-8 tw-mb-4">
       {data?.results?.map((item, idx) => {
         return <BookCard key={idx} data={item} />;
       })}