+++ src/services/bussiness/router/types.ts
@@ -10,16 +10,16 @@ export const ROUTER_NAME = {
   DEV_DETAIL: 'devDetail',
 } as const;
 
-export type RouterNameType = 
-| typeof ROUTER_NAME.ABOUT
-| typeof ROUTER_NAME.BOOK
-| typeof ROUTER_NAME.CHORE
-| typeof ROUTER_NAME.DEV
-| typeof ROUTER_NAME.HOME
-| typeof ROUTER_NAME.BOOK_DETAIL
-| typeof ROUTER_NAME.CHORE_DETAIL
-| typeof ROUTER_NAME.CONFIDE_DETAIL
-| typeof ROUTER_NAME.DEV_DETAIL;
+export type RouterNameType =
+  | typeof ROUTER_NAME.ABOUT
+  | typeof ROUTER_NAME.BOOK
+  | typeof ROUTER_NAME.CHORE
+  | typeof ROUTER_NAME.DEV
+  | typeof ROUTER_NAME.HOME
+  | typeof ROUTER_NAME.BOOK_DETAIL
+  | typeof ROUTER_NAME.CHORE_DETAIL
+  | typeof ROUTER_NAME.CONFIDE_DETAIL
+  | typeof ROUTER_NAME.DEV_DETAIL;
 
 export const ROUTER_PATH_NAME = {
   [ROUTER_NAME.HOME]: '/',
@@ -30,5 +30,5 @@ export const ROUTER_PATH_NAME = {
   [ROUTER_NAME.BOOK_DETAIL]: '/book/[id]',
   [ROUTER_NAME.CHORE_DETAIL]: '/chore/[id]',
   [ROUTER_NAME.CONFIDE_DETAIL]: '/confide/[id]',
-  [ROUTER_NAME.DEV_DETAIL]: '/dev/[id]'
-}
+  [ROUTER_NAME.DEV_DETAIL]: '/dev/[id]',
+};