+++ tailwind.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  prefix: 'tw-',
   content: [
     "./src/pages/**/*.{js,ts,jsx,tsx}",
     "./src/components/**/*.{js,ts,jsx,tsx}",